+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to
 
 - Changed how `j1-integration visualize` displays placeholder entities. Now only
   properties present in `targetFilterKeys` are displayed in the graph, making
-  target entities smaller.
+  target entities smaller. Also, set borders of placeholder entities to dashed.
 - Bump `@jupiterone/data-model` to expose `Alert` entity schema.
 - `createEventPublishingQueue` takes in an optional Axios config.